dom.ipc.processCount is by default set to 8 in FF 66
This is (imho) still to low, increase it to 10 since 8GB+ RAM is already standard for several years now.

http://www.erahm.org/2019/03/13/doubling-the-number-of-content-processes-in-firefox/
https://bugzilla.mozilla.org/show_bug.cgi?id=1470280

layout.css.servo.enable is enabled for e.g. Waterfox. 
https://bugzilla.mozilla.org/show_bug.cgi?id=1294570

It can result together with a higher processCount in overall better performance at the risk to break some extensions, I tested it and nothing broke for me.