Re-added pref.privacy.disable_button.view_passwords
The pref does not work on newer FF (it's just to maintain backward compatibility to older FF versions).

The new "thing" is to work with "toolkit.legacyUserProfileCustomizations.stylesheets", true instead (we enabled it, remember?)

From there we can create (manually) in our Chrome.css:
- [active firefox profile]\chrome\userChrome.css
- button#togglePasswords{display:none !important}

This will now truly hide the password button (which we want to hide) because we use KeePass and other pass. manager.