Guard added
It's currently beta: https://useguard.com/blog/future/.
Website: https://useguard.com/

It's beta, that been sad it has stuff to learn because currently it ranks Mozilla on par with YouTube (as rank C) while e.g. Linked In gets B. This is, of course, a AI problem which needs to be reported/resolved.

However the program/AI/website might be useful to quickly check the privacy policy of website/app X and that's a reason to mention it.