Workaround: Compressing the profile database to "gain some speed"
This is basically the main reason what people complain about since years, it's not disk cache it's your "slow database" filled with some garbage. Firefox by default vacuums the database every 30 days - compared to disk cache 14-15 days. So compressing the database will speedup Firefox. However, there are no specific flags for it and that's the reason why it's listed under workarounds. I guess there is no flag for it because it's possible that you loose some settings/extensions, so make a backup before you vacuum your entire profile. A more secure way is to e.g. use CCleaner, but it only vacuum one specific database instead of the whole profile.