Webassembly (Wasm) got official approved by World Wide Web Consortium (W3C)
Press:
https://www.w3.org/2019/12/pressrelease-wasm-rec.html.en

Specs:
https://www.w3.org/TR/wasm-core-1/

Web API:
https://www.w3.org/TR/wasm-web-api/

JS Interface:
https://www.w3.org/TR/wasm-js-api/

That been said, now that this is "official" it opens fixes and patches for every browser, I'm sure the known "problems" are solved within a few weeks. Wasm was designed with security in mind, and we should not cripple each new API, all is fixable.