Ref: Cloudflare possible interercept or even decrypt TLS
It needs to be reviewed but the decryption of TLS part is true. However, CLoudflare already documents it and says that this is ony temporarily and that the logs are been deleted. 

I see this much more as a Cloudflare privacy concern rather than a Mozilla issue. Mozilla is just the partner and features like DOH are still beta.