Enable Canvas 2D
Currently Nightly only.

Overview:
The latest Nightly has the pref re-enabled to move accelerated Canvas 2D to the GPU process on Windows.

This is only enabled on Nightly at the moment and is part of our work to move GPU access and win32k system calls out of the content process.

Clownzilla:
https://bugzilla.mozilla.org/show_bug.cgi?id=1548487